Add interface options and "all" to map popup.
Allow a user to choose all nodes or a specific node. Allow a user to
choose all interfaces or a specific one. Control the drop down menus
based on the user choices. If cpu or memory is selected, disable the
interfaces drop down. Populate the interfaces drop down based on the
selected node.